Update info for global rules
Realized that the example provided in the Proxy GitHub is actually incorrect (just filed a PR to correct that). The way "global" rules work is that they apply to all explicitly specified ports in the preprocessor rule file. Additionally, the global rules must be at the end of the file (since it looks for all previously specified ports in the file and so, if the global rule was at the beginning, there would be no other rules above/previous to it). Hope that makes sense